FORMS-12203: Enhanced test cases
